Solution2 (py-scrape-quotes) with separate file for settings
